+++ lib/pay/stripe/charge.rb
@@ -104,7 +104,7 @@ module Pay
       # refund!(5_00, refund_application_fee: true)
       def refund!(amount_to_refund, **options)
         ::Stripe::Refund.create(options.merge(charge: processor_id, amount: amount_to_refund), stripe_options)
-        pay_charge.update(amount_refunded: amount_to_refund)
+        pay_charge.update(amount_refunded: pay_charge.amount_refunded + amount_to_refund)
       rescue ::Stripe::StripeError => e
         raise Pay::Stripe::Error, e
       end